// codegen/CwAsteroid.client.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@1.11.1.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

import { CosmWasmClient, SigningCosmWasmClient, ExecuteResult } from "@cosmjs/cosmwasm-stargate";
import { Coin, StdFee } from "@cosmjs/amino";
import { Addr, InstantiateMsg, ExecuteMsg, ExecuteMsg1, Uint128, GameConfig, QueryMsg, GameRoundStatus, GameRound, GameStatus, Game, Uint64, Leaderboard } from "./CwAsteroid.types";
export interface CwAsteroidReadOnlyInterface {
  contractAddress: string;
  getGame: ({
    gameId
  }: {
    gameId: number;
  }) => Promise<Game>;
  getLeaderboard: () => Promise<Leaderboard>;
  getCurrentRound: ({
    gameId
  }: {
    gameId: number;
  }) => Promise<GameRound>;
  getGameStatus: ({
    gameId
  }: {
    gameId: number;
  }) => Promise<GameStatus>;
  getGamesCount: () => Promise<Uint64>;
}
export class CwAsteroidQueryClient implements CwAsteroidReadOnlyInterface {
  client: CosmWasmClient;
  contractAddress: string;
  constructor(client: CosmWasmClient, contractAddress: string) {
    this.client = client;
    this.contractAddress = contractAddress;
    this.getGame = this.getGame.bind(this);
    this.getLeaderboard = this.getLeaderboard.bind(this);
    this.getCurrentRound = this.getCurrentRound.bind(this);
    this.getGameStatus = this.getGameStatus.bind(this);
    this.getGamesCount = this.getGamesCount.bind(this);
  }
  getGame = async ({
    gameId
  }: {
    gameId: number;
  }): Promise<Game> => {
    return this.client.queryContractSmart(this.contractAddress, {
      get_game: {
        game_id: gameId
      }
    });
  };
  getLeaderboard = async (): Promise<Leaderboard> => {
    return this.client.queryContractSmart(this.contractAddress, {
      get_leaderboard: {}
    });
  };
  getCurrentRound = async ({
    gameId
  }: {
    gameId: number;
  }): Promise<GameRound> => {
    return this.client.queryContractSmart(this.contractAddress, {
      get_current_round: {
        game_id: gameId
      }
    });
  };
  getGameStatus = async ({
    gameId
  }: {
    gameId: number;
  }): Promise<GameStatus> => {
    return this.client.queryContractSmart(this.contractAddress, {
      get_game_status: {
        game_id: gameId
      }
    });
  };
  getGamesCount = async (): Promise<Uint64> => {
    return this.client.queryContractSmart(this.contractAddress, {
      get_games_count: {}
    });
  };
}
export interface CwAsteroidInterface extends CwAsteroidReadOnlyInterface {
  contractAddress: string;
  sender: string;
  lifecycle: (executeMsg: ExecuteMsg, fee?: number | StdFee | "auto", memo?: string, _funds?: Coin[]) => Promise<ExecuteResult>;
}
export class CwAsteroidClient extends CwAsteroidQueryClient implements CwAsteroidInterface {
  client: SigningCosmWasmClient;
  sender: string;
  contractAddress: string;
  constructor(client: SigningCosmWasmClient, sender: string, contractAddress: string) {
    super(client, contractAddress);
    this.client = client;
    this.sender = sender;
    this.contractAddress = contractAddress;
    this.lifecycle = this.lifecycle.bind(this);
  }
  lifecycle = async (executeMsg: ExecuteMsg, fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    return await this.client.execute(this.sender, this.contractAddress, {
      lifecycle: executeMsg
    }, fee, memo, _funds);
  };
}